[story/VSPC-52] fixed indentation and raw type warnings.

+++ vspace/src/main/webapp/WEB-INF/views/staff/modules/slides/slide.jsp
@@ -179,22 +179,31 @@ $(document).ready(function() {
 		e.preventDefault();
 		$("#confirmDeleteImageAlert").hide();
 		var blockId = $('#deleteImageId').attr('value');
-		$('#deleteImageId').remove()
-		// ------------- delete image content blocks ------------
-		$.ajax({
-		    url: "<c:url value="/staff/module/${module.id}/slide/${slide.id}/image/" />" + blockId + "?${_csrf.parameterName}=${_csrf.token}",
-		    type: 'DELETE',
-		    success: function(result) {
-		        $('#' + blockId).remove();
-		    },
-			error: function(data) {
-				var alert = $('<div class="alert alert-danger alert-dismissible fade show" role="alert"><p>We are sorry but something went wrong. Please try to delete again later.</p><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
-				$('.error').append(alert); 
-				$(".error").delay(4000).slideUp(500, function(){
-				    $(".error").empty();
-				});
-			}
-		});
+		console.log(blockId);
+		if(blockId == null || blockId == ''){
+			var alert = $('<div class="alert alert-danger alert-dismissible fade show" role="alert"><p>We are sorry but something went wrong. Please try to delete again later.</p><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+			$('.error').append(alert); 
+			$(".error").delay(4000).slideUp(500, function(){
+			    $(".error").empty();
+			});
+		} else{
+			$('#deleteImageId').remove()
+			// ------------- delete image content blocks ------------
+			$.ajax({
+			    url: "<c:url value="/staff/module/${module.id}/slide/${slide.id}/image/" />" + blockId + "?${_csrf.parameterName}=${_csrf.token}",
+			    type: 'DELETE',
+			    success: function(result) {
+			        $('#' + blockId).remove();
+			    },
+				error: function(data) {
+					var alert = $('<div class="alert alert-danger alert-dismissible fade show" role="alert"><p>We are sorry but something went wrong. Please try to delete again later.</p><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
+					$('.error').append(alert); 
+					$(".error").delay(4000).slideUp(500, function(){
+					    $(".error").empty();
+					});
+				}
+			});
+		}
 	});
 	
 	$("#addImgAlert").draggable();
